Add support for Ubuntu 24.04
Updated the gym version from 0.19.0 -> 0.25.2
gym versions prior to 0.22.0 fail to build on Ubuntu 24.04
gym version 0.22.0 introduced a bug that didn't accept the package:environment syntax for gym.make(), which was corrected in version 0.24.0
According to the release notes, versions 0.24.0 and 0.24.1 have other significant bugs and should not be used.
Thus, version 0.25.0 is the first version we can use. Versions 0.25.1 and 0.25.2 are simply bug fixes to 0.25.0.
Version 0.25.2 works well and is the most recent version that doesn't require an API change. It does produce warnings of "future" API changes, though.